bug: parser drops last line in each chunk
Each chunk processed is not including the EOL so those last lines in
those chunks are being dropped. Fix by including the last byte after
searching for EOL.

Also, files that don't end in eol aren't getting their last line parsed.  Fix this by
pretending there is an EOL after the last byte when the last byte is not EOL.