use relative scrolling/painting instead of refresh whole screen
When part of the buffer is still on the screen from the previous refresh,
just scroll that part into view and repaint the updated lines.  This seems overly
complicated, and maybe I can simplify it later. But it also has a bug.

When the screen is resized larger than before, the new space doesn't
get painted correctly.  This problem persists forever.