Simplify Phrase and use linear search to handle inserts
Track phrases with just start and pattern because length can be inferred from the
next phrase in the sentence.  This simplifies the logic around deciding where to
insert the phrase, what it overlaps, and the work needed to accomplish it.

But the logic to handle the various possible insertion conditions seems
convoluted.  :-\