Location: treat virtual locations as "edge" locations
Previously I treated these as actual values.  I.e., StartOfFile == FirstLine.
But that makes Index(0,0) ambiguous and a special case.

Make the virtual locations into edges.  We use them to specify a start
location, and we use them to detect running into the edge. But
we do not treat them as actual lines.