Add target offset to gaps to show what we wanted
When we resolve gaps it helps to know what target we wanted to load next.
Put the "missed" target offset into all gaps. Fix up `index_chunk` to
use the target to find which chunk to load.

Future: find the chunk_size by querying the underlying file impl. This
will help when we want to load zstd frames, for instance.