Impl BufRead for AsyncStdin
AsyncStdin is already completely buffered, so it is unnecessary to also
wrap it in a BufReader.  We can implement the BufRead trait trivially.

Do that, and clean up Seek a bit.

Expect in the future that all LogFile providers can provide a
{Seek, BufRead} object generically.

Note this doesn't help us with filtered lines.  There's no way to implement
BufRead.lines() such that it uses our internal indexes.  No biggie.