impl ReadBuf for CompressedFile
Add a ReadBuffer to manage the buffer (a Vec<u8>) and our current position
within it.  Try to keep it bounded to a reasonable size.  When we read contiguously
from the compressed file, extend the buffer.  When seeking to a new frame,
invalidate the buffer.  When seeking within the buffer's held data, avoid
re-reading compressed data again.

TODO: Add tests that prove we can walk the lines backwards without
re-reading all the frames for every line.