CachedStream: wait for data in Read -- again
Fixed this before for Read() buf not for BufRead().  Change fill_buffer()
to do nothing if the caller's target plus some threshold is already reached, to try to read from the queue if the threshold is not reached, and to block if the caller's target position is past current eof. Once we decide to read, continue to drain the queue until it is empty, the file is closed, or the threshold is reached.  The threshold here represents our
push-back on the source stream that stops the sender from providing
more work that we're not yet ready to receive.

Move it all into common code for clarity and add some helper functions
to wrap the individual pieces.