index gaps carry target offset and direction
When we try to find a target but encounter a gap, we need to index the next
chunk of data to resolve our target.  But our target is imprecise since we don't
have the indexed chunk yet.  When we do index it and try to land on our target
it's possibly in the middle of some line. Do we want the start of the line containing
the offset, or do we want the next line next?

Well, it depends.  Are we stepping forward or backward through the file? By the
time the gap is resolved, we've forgotten which way we wanted to go. We
need this information to support reverse iterators.

Extend the TargetOffset to hold a direction with its offset: After or AtOrBefore.
Then seek the direction indicated when we ultimately get to resolve the
location after merging the gap.

Also impl some helpers to our Enums to extract info more directly.

Implement Eq/PartialEq for Location values so DoubleEndedIterator can
determine when the back meets the front.

Fix iterators to use new TargetOffset and changed API.

Fix a bug in chunk reading that caused us to read too many bytes.