EventualIndex: Handle stale locations automatically
I was considering making Location a proper cursor that owns the IndexedFile.
This would solve some issues, such as the fact that a Location can become
stale if the EventualIndex grows.  But it would also mean that I can't have
two Locations open on a file at once, such as a DoubleEndedIterator.

Wait -- I have one of those. Do I have a problem with a location becoming
invalid when iterating out from the middle, for instance? Yes, it turns out
I do.  Add the test_iterator_middle_out to demonstrate it.

Handle stale cursors by binary-searching the indexes to find the correct
location for a given offset (and the line before or after it).