move iterators into Log to return LogLine
Log as a simple type wrapper can't hold the TimeStamper we need to fill
out LogLines. Make it into a proper struct so it can hold a unique TimeStamper
for each LineIndexer<LogSource>. We need that because each log source
might have a different time stamp format, and we hope to learn it in the
future to optimize the conversion time per file.

Give it a conversion from LogSource and LogBase to ease the transition
for its tests and callers.