CachedStream: quench file during init for tests
Kind of a hack... Cached stream is empty when we first try to iterate,
so a rev-iterator will iterate from the start of the file (zero length).
We can work around that if the file is actually all available by quenching
first.  But we should probably try to quench() before we first try to
iterate from Virtual::End, right?

Let's use this quick fix for the tests. If we don't do it for the real code,
it's probably ok since the caller should be looking for more data once
he goes idle anyway.  Maybe.