renamed VirtualLocation::After => AtOrAfter
That better reflects what it does because it always includes the given offset in
the search for the next position.  But since it was named wrong, we sometimes
exclude the current position in comparisons (which is a bug).  Sad that no test
noticed.  But the new name is easier to reason about.