Attempt to simplify filtered-log reading by using iterators
Filtered-log should be able to do all of its gap-filling by iterating on its internal Log
only when needed and filling the gaps as it goes.  But the code is too complicated now,
and it relies on complicated Location traversal between the two interfaces.

Attempt to fix this by iterating the internal log instead of leaning on resolve_location.
Unfortunately, we can't call the iterate function on a trait, so this also means we have
to abandon polymorphic generics and implement monomorphic ones instead. This
was easier than expected, but I fear it will constrain future code too much.  But
it works at the moment.

Possible path forward:
1. isolate the Location enums for public consumption from the ones used internally
   to help simplify the gap shenanigans.
2. Introduce a new interface similar to iteration that we can still use with traits (one
   that does not require Sized).
3. Simplify the filtered_log gap-filler to pass-the-buck on iterators more cleanly.
4. Revert to using Box<dyn IndexedLog> for FilteredLog::log