add &-filter prompt and function
And it works. woot!

Currently it is blocking while it searches.
Search for non-matching term just hangs while the search churns.
    On my 500M test file,
      grok: takes 23 seconds
      less: takes 4.5 minutes

I saw some filter bug when paging around a file with filtering on and we hit some assert.
I didn't save the output. I expect it won't be hard to repro.