grok: adjust timeouts for keys and backfilling
Background fill is slow when we don't give it much CPU. Adjust the
keyboard event loop to timeout immediately if we have back-filling to
do.  Increase the back-fill timer to give it more chance to catch up, too.

40ms seems like a reasonable delay between key events, sometimes,
but we can do better by adjusting to less time when someone is actively
typing or mousing. Add a FIXME to address this later, maybe.