fixed up sane_index and made conforming iterators
Make iterators conform to Rust specs.  That is, rev iterator stops
when it meets fwd iterator.  Also, seeking from within any line will
start the iterator on the line containing that offset.  This is essential
for sub_line_iterator to make sense.  Otherwise we would not be able
to iterate from a sub-line chunk in the file in a reasonable way.

All tests pass.

Still need to fix up background resolver and some todo!'s.