Add end-pos to Mapped waypoints
We don't need it to find the line ending, but we can use it to find the
line containing our target.

Note that this makes all the waypoint index contiguous for the base
indexed file.   Filtered indexes will still have gaps.  But we'll be able to
tell which line our offset touches when searching.