Fix filtered/inner log position sync bug
When stepping through gaps, we forgot to use the current offset and always used the original
one instead.  This confused the gap resolver since it received lines out of range of the gap.
We need the original offset for the first step which may land in the
middle of a gap.  Thereafter, we need the "next" offset.