Avoid unnecessary seeks
Seeking to the start of each line is my performance killer.  I can avoid
actually seeking by tracking the current position and skipping the
seek when we're already in the right place.

The big issue is that bufread always resets the position without
considering whether it's already the current position.  So it always
resets the buffer, and this negates the whole benefit of bufread.

Note this trick is already implemented in CompressedFiles from
long ago.