sub_line_indexer: Fix overlapping fwd/next
next / next_back no longer allows a line to be split up across different
directions.  This is caused by several bugs:
 - consumed is not set correctly after a line is fully consumed
 - return value aborts when iterators are equal, but it should go one more
 - We quit when source iterator quits, even though we still may have
   some remaining line to hand out.

Fix these by copying the line from the other iterator when we run out, if
it has any.  In practice this only affects the rev() iterator, but it's possible
for both to get called independently as we do in a test.

Fix the calculation of the consumed marker by setting the index correctly
after consuming the whole line, and before calling mark_consumed.

Fix the <= comparison to be < instead. This way the iterators detect
when they have passed each other, which is useful since they are
updated to the part consumed, but they are pointing at opposite ends.

Alse, clear the start field after we use it, just so we don't have to keep
checking it later on, and to make debugging easier.

Re-enable a test that exercises this and proves it works.