FilteredLog: refactor filter into LogFilter
I need to be able to isolate the log from the filter, I think.  Anyway,
I will have multiple filters, and I need them to work with the same source
IndexedLog or something like that.  Isolate the LogFilter part of the
old FilteredLog so it takes Log as a local argument each time.