GetLine's pos should reference the found pos, not the next one
We advance the position to point the next place to search inside IndexedLog::next
and next_back as a convenience. This is because we assume the caller will want to
start their search there on the next call. (Before this, we used to advance the pos
at the start of any search, but this caused us to skip over line 1 sometimes.)

Now the problem is that when we want to find the next filtered line position, we
may fail because next() returns the next unmapped position after our actually-found
line.

When we jump to a position in the middle of the file, for example, we don't find
any lines if we rely on "pos.is_mapped()".  And when we search for a filtered next
occurrence, we also think we didn't find it when it returns "unmapped".

Additionally, the GetLine semantics are simply inconsistent since many internal
funcions use the "found-line" position, and we convert them to "next-line" pos
at some critical points in the code.  Let's make this consistent.

Separate the "find_next" and "advance_position" functions in IndexedLog and
change the callers to deal with the new reality.