erase_back: fix bug when resolving last bit of the first gap
We try to step backwards if the `erase` ended up splitting a gap in two.  But we may accidentally
step forwards also if the gap was completely resolved and it happens to be the first row in the
index.  This works ok, but there's an assert that is too sensitive.

Allow the position to become "invalid" if we try to step left and walk off the beginning of the index.