Isolate filtered log in log_stack to appease Rust
Move the Filtered view of the log out into its own struct so we
don't confuse the Borrow checker when we want to search
inside the filtered log.

Also, make search look at the filtered-log view instead of the raw
log.  Reset the search index when the filter changes.

Also, add index.reset().