filtered next_back confused position with inner position
When Filter tries to step backwards and finds no match in a gap, it returns the inner position instead
of the correct position for the filtered index.  This causes expensive lookups and wrong logic.  Fix it
to return the correct results with the correct filtered-index position instead.

Also correct the check for BOF to check the indexes instead of checking if the position didn't change.
It turns out the position might not change in several Miss scenarios.   (Can it, though?)