Track addressable-space in index
Maybe we need to know how many bytes are possible mappable
in the index.  This is for the search over a filtereed log.  We don't find
lines from the whole file; just the visible part of the file. Try to stop
searching when we've seen everything we can.

WIP: Probably the wrong idea.  Going to revert this next.  Saving it for posterity.