Erase skipped regions from indexes
When we are building an index using a filtered log, there will be lines that are skipped
by the filter. Currently we leave these as unexplored gaps in our index on the assumption
that they may be useful to us in the future.  But this is impractical because we will discard
our index if the filter changes later, and it causes us to think our index is incomplete
because it still has gaps.

Clear the gaps when we see them while walking through the filter.