check if inner_pos is out of sync and recreate it if so
I don't know if this actually happens, but I realized it can while
running some bad code against .zst files and hitting an unexpected
assert.  But realistically, if we're resolving gaps and the file is partially
resolved separately, then inner_pos may have "moved" since we
last looked.  Check to see if it moved and rebuild it if so.