Update crossterm to 28.1
I thought this would update automatically, but it doesn't.  Maybe some cargo command
is needed.  Ran into incompatibility with reedline and needed to update this.  But it
broke lots of crossterm::Result references.  Not sure why.  Can't find it mentioned in the
release notes.   But this seems to work for now.