Separate styling from sanitizing and rework the callers
We want to apply all styles to the raw line before we sanitize it. Otherwise
we cannot search it properly.  But we also need to stretch the style positions
after we sanitize the line.

Add a regex to identify the pieces we will sanitize.  Then sanitize
them after all the other styles are applied.  Finally, adjust the style
positions to align with the new sanitized line.

Also, remove obsolete style code from Document.