Fix scroll tracking bug on short files
When a file has fewer lines than the display height, the displayed_lines
tracker gets confused.  It doesn't get truncated to the actual displayed
lines.  This causes weird scrolling bugs as we never know what line is at
the top of the display.

Fix this to truncate to the size of the actual lines on the screen.