Don't busy-loop if the log source is empty
We continuously re-init the screen when there are no lines to log. Avoid this by dropping the
unneeded initial-paint and handling the "no displayed lines" condition by just using zero for
the start and end offsets in this case.

Testing: cat /dev/null | cargo run --bin grok

Also noticed that we don't update partially-retrieved lines later on, so add a bug report.