stream: reading from a small partial last-line causes poll to hang
When we read a partial last-line (one with no LF at the end) poll gets stuck on the next call
as it thinks there is no new data and it wants to wait for it to satisfy slow streams.  Give poll
a timeout and let it "finish" when it sees that the stream is not making progress.

This is a bit of a hack and the timeout is fixed / reused each time. But it softens this edge
case so let's run with it for now.