compressed_file: always stream bytes to end of current buffer
The code to figure out where to put new incoming bytes was borked.
It took the current read_pos as the location for the new bytes, but they
should always be appended to the existing read_buffer.  If they're not
contiguous, the external caller is responsible for initializing a new
buffer.  So remove the "position" idea and the buffer-reset function
from the buffer::extend function.  Require new buffers to come with
the position of the expected first byte so they can track from there on.

There seems to still be a bug somewhere.  It can be repro'ed with
this sequence:

1. cargo run --bin grok --release -- -X --color ./indexed_file/resources/test/core.log-2022040423.zst
2. Filter on "segmap.*flush"
3. Search for "123" (after some significant amount of the filter is resolved).

We get an assert that seems to show that some line is being returned
which is not aligned with a remaining gap.  This shouldn't happen, so
maybe we're still getting borked buffers somewhere.  Or maybe the
LogFilter code still has a bug.