Add rudimentary gzip support
This implementation decompresses the gzip into a memory cache.  It's reasonably fast
but it eats a lot of memory instead of decompressing as we go.

Bug?  It seems to stop scanning the cache if the reader got an error. For example, if we
have a truncated gz file, one that complains about "unexpected end of file" when
decompressing, then the cached reader won't background resolve gaps.  wtf?